requirements.txt
fastapi
uvicorn
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings
pyjwt
psycopg2-binary
python-dotenv
alembic